refactor: #50 StudyHeaderTitle Block 이름 변경
