+++ src/components/pages/Study/StudyHeader/StudyName/StudyName.tsx
